Модуль utils разделён на 3 модуля
